hotfix on kegg pathway to add back the up DGE

+++ sequana/enrichment/kegg.py
@@ -184,13 +184,12 @@ class KEGGPathwayEnrichment:
 
             pathways = glob.glob(preload_directory + "/*json")
             logger.info(f"Loading {len(pathways)} pathways from local files in {preload_directory}.")
-            for i, name in enumerate(pathways):
+            for name in tqdm(pathways):
                 key = name.strip(".json").split("/")[-1]
                 with open(name, "r") as fin:
                     data = json.load(fin)
                     self.pathways[key] = data
         else:  # pragma: no cover  #not tested due to slow call
-
             for ID in tqdm(self.kegg.pathwayIds, desc="Downloading KEGG pathways"):
                 self.pathways[ID.replace("path:", "")] = self.kegg.parse(self.kegg.get(ID))
 
@@ -377,7 +376,8 @@ class KEGGPathwayEnrichment:
     def _get_summary_pathway(self, pathway_ID, df, l2fc=0, padj=0.05):
 
         genes = self.df_pathways.loc[pathway_ID]["GENE"]
-        df_down = df.query("padj<=@padj and log2FoldChange<@l2fc").copy()
+
+        df_down = df.query("padj<=@padj and log2FoldChange<=-@l2fc").copy()
         df_up = df.query("padj<=@padj and log2FoldChange>=@l2fc").copy()
 
 
@@ -385,7 +385,7 @@ class KEGGPathwayEnrichment:
         # since names is suppose to have been populated with the annotaion, it should be
         # found. no need for sanity checks.
         df_down["Name"] = df_down[self.color_node_with_annotation]
-        df_up["Name"] = df_down[self.color_node_with_annotation]
+        df_up["Name"] = df_up[self.color_node_with_annotation]
 
 
         # in principle, the KEGG pathays field 'GENE' is stored as 
@@ -505,11 +505,9 @@ class KEGGPathwayEnrichment:
         summary = self._get_summary_pathway(pathway_ID, df)
         colors = self._get_colors(summary)
 
-        logger.info(f"pathway {pathway_ID} total genes: {len(summary)}")
         count_up = len(summary.query("type == '+'"))
         count_down = len(summary.query("type == '-'"))
-        logger.info(f"this pathway down-regulated genes: {count_down}")
-        logger.info(f"this pathway up-regulated genes: {count_up}")
+        logger.info(f"pathway {pathway_ID} total genes: {len(summary)}. Found {count_down} down-regulated and {count_up} up-regulated")
 
         url = "https://www.kegg.jp/kegg-bin/show_pathway"
         # dcolor = "white"  --> does not work with the post requests unlike get
@@ -563,47 +561,27 @@ class KEGGPathwayEnrichment:
 
         return summaries
 
-    def find_pathways_by_gene(self, gene_name, match="exact"):
+    def find_pathways_by_gene(self, gene_name):
         """Returns pathways that contain the gene name
 
-        ke.find_pathways_by_gene("ysgA")
+        ::
+
+            ke.find_pathways_by_gene("ysgA")
+
         """
 
-        # First let us find the kegg ID
-        genes = self.kegg.list(self.kegg.organism).strip().split("\n")
+        # we are going to search for the gene name and lower case onyl
+        candidates = [gene_name, gene_name.lower()]
 
-        keggid = [x.split("\t")[0].strip() for x in genes]
-        gene_names = [x.split("\t")[1].split(";")[0].strip() for x in genes]
+        found_paths = []
 
-        self.keggid = keggid
-        self.gene_names = gene_names
-        candidates = []
-        for x, y in zip(keggid, gene_names):
-
-            if match == "exact":
-                if gene_name == y:
-                    candidates = x.split(":")[1]
-                    break
-            else:
-                if gene_name in y:
-                    candidates.append(x)
-        if match != "exact":
-            candidates = [x.split(":")[1] for x in candidates]
-            logger.info("Found {} candidate(s): {}".format(len(candidates), candidates))
-        else:
-            logger.info("Found {} in {}".format(gene_name, candidates))
-
-        paths = []
         for key in self.pathways.keys():
             if "GENE" in self.pathways[key]:
-                if match == "exact":
-                    if candidates in self.pathways[key]["GENE"].keys():
-                        paths.append(key)
-                else:
-                    for candidate in candidates:
-                        if candidate in self.pathways[key]["GENE"].keys():
-                            paths.append(key)
-        return list(set(paths))
+                genes = set([x.split(";")[0] for x in self.pathways[key]["GENE"].values())
+                for candidate in candidates:
+                    if candidate in genes:
+                        found_paths.append(key)
+        return list(set(found_paths))
 
     def save_project(self, tag, outdir="."):
         """Save tables and visualisations of the complete enrichmment analysis."""
@@ -642,5 +620,7 @@ class KEGGPathwayEnrichment:
         os.makedirs(outdir, exist_ok=True)
 
         for key, data in self.pathways.items():
-            with open(f"{outdir}/{key}.json", "w") as fout:
-                json.dump(data, fout)
+            output_json = f"{outdir}/{key}.json"
+            if not os.path.exists(output_json):
+                with open(output_json, "w") as fout:
+                    json.dump(data, fout)
